+++ Tests/TuistGeneratorTests/Linter/GraphLinterTests.swift
@@ -247,4 +247,119 @@ final class GraphLinterTests: XCTestCase {
         // Then
         XCTAssertFalse(result.isEmpty)
     }
+
+    func test_lint_missingProjectConfigurationsFromDependencyProjects() throws {
+        // Given
+        let customConfigurations: [BuildConfiguration: Configuration?] = [
+            .debug("Debug"): nil,
+            .debug("Testing"): nil,
+            .release("Beta"): nil,
+            .release("Release"): nil,
+        ]
+        let targetA = Target.empty(name: "TargetA", product: .framework)
+        let projectA = Project.empty(path: "/path/to/a", name: "ProjectA", settings: Settings(configurations: customConfigurations))
+
+        let targetB = Target.empty(name: "TargetB", product: .framework)
+        let projectB = Project.empty(path: "/path/to/b", name: "ProjectB", settings: Settings(configurations: customConfigurations))
+
+        let targetC = Target.empty(name: "TargetC", product: .framework)
+        let projectC = Project.empty(path: "/path/to/c", name: "ProjectC", settings: .default)
+
+        let graph = Graph.create(projects: [projectA, projectB, projectC],
+                                 dependencies: [
+                                     (project: projectA, target: targetA, dependencies: [targetB]),
+                                     (project: projectB, target: targetB, dependencies: [targetC]),
+                                     (project: projectC, target: targetC, dependencies: []),
+                                 ])
+
+        // When
+        let result = subject.lint(graph: graph)
+
+        // Then
+        XCTAssertEqual(result, [
+            LintingIssue(reason: "The project 'ProjectC' has missing or mismatching configurations. It has [Debug (debug), Release (release)], other projects have [Beta (release), Debug (debug), Release (release), Testing (debug)]",
+                         severity: .warning),
+        ])
+    }
+
+    func test_lint_mismatchingProjectConfigurationsFromDependencyProjects() throws {
+        // Given
+        let customConfigurations: [BuildConfiguration: Configuration?] = [
+            .debug("Debug"): nil,
+            .debug("Testing"): nil,
+            .release("Beta"): nil,
+            .release("Release"): nil,
+        ]
+        let targetA = Target.empty(name: "TargetA", product: .framework)
+        let projectA = Project.empty(path: "/path/to/a", name: "ProjectA", settings: Settings(configurations: customConfigurations))
+
+        let targetB = Target.empty(name: "TargetB", product: .framework)
+        let projectB = Project.empty(path: "/path/to/b", name: "ProjectB", settings: Settings(configurations: customConfigurations))
+
+        let mismatchingConfigurations: [BuildConfiguration: Configuration?] = [
+            .release("Debug"): nil,
+            .release("Testing"): nil,
+            .release("Beta"): nil,
+            .release("Release"): nil,
+        ]
+        let targetC = Target.empty(name: "TargetC", product: .framework)
+        let projectC = Project.empty(path: "/path/to/c", name: "ProjectC", settings: Settings(configurations: mismatchingConfigurations))
+
+        let graph = Graph.create(projects: [projectA, projectB, projectC],
+                                 entryNodes: [targetA],
+                                 dependencies: [
+                                     (project: projectA, target: targetA, dependencies: [targetB]),
+                                     (project: projectB, target: targetB, dependencies: [targetC]),
+                                     (project: projectC, target: targetC, dependencies: []),
+                                 ])
+
+        // When
+        let result = subject.lint(graph: graph)
+
+        // Then
+        XCTAssertEqual(result, [
+            LintingIssue(reason: "The project 'ProjectC' has missing or mismatching configurations. It has [Beta (release), Debug (release), Release (release), Testing (release)], other projects have [Beta (release), Debug (debug), Release (release), Testing (debug)]",
+                         severity: .warning),
+        ])
+    }
+
+    func test_lint_doesNotFlagDependenciesWithExtraConfigurations() throws {
+        // Lower level dependencies could be shared by projects in different workspaces as such
+        // it is ok for them to contain more configurations than the entry node projects
+
+        // Given
+        let customConfigurations: [BuildConfiguration: Configuration?] = [
+            .debug("Debug"): nil,
+            .release("Beta"): nil,
+            .release("Release"): nil,
+        ]
+        let targetA = Target.empty(name: "TargetA", product: .framework)
+        let projectA = Project.empty(path: "/path/to/a", name: "ProjectA", settings: Settings(configurations: customConfigurations))
+
+        let targetB = Target.empty(name: "TargetB", product: .framework)
+        let projectB = Project.empty(path: "/path/to/b", name: "ProjectB", settings: Settings(configurations: customConfigurations))
+
+        let additionalConfigurations: [BuildConfiguration: Configuration?] = [
+            .debug("Debug"): nil,
+            .debug("Testing"): nil,
+            .release("Beta"): nil,
+            .release("Release"): nil,
+        ]
+        let targetC = Target.empty(name: "TargetC", product: .framework)
+        let projectC = Project.empty(path: "/path/to/c", name: "ProjectC", settings: Settings(configurations: additionalConfigurations))
+
+        let graph = Graph.create(projects: [projectA, projectB, projectC],
+                                 entryNodes: [targetA],
+                                 dependencies: [
+                                     (project: projectA, target: targetA, dependencies: [targetB]),
+                                     (project: projectB, target: targetB, dependencies: [targetC]),
+                                     (project: projectC, target: targetC, dependencies: []),
+                                 ])
+
+        // When
+        let result = subject.lint(graph: graph)
+
+        // Then
+        XCTAssertEqual(result, [])
+    }
 }